requirements.txt
fastapi
uvicorn
boto3
sqlalchemy
alembic
python-dotenv
psycopg2-binary
gunicorn
prometheus-fastapi-instrumentator
fastapi-cache2